requirements.txt
Django~=3.2.7
mysqlclient~=2.0.3
pymysql~=1.0.2
pyjwt~=2.1.0